Starting new milestones:
Dockerizing mcp server. Then dockerizing apps.

Why?

Because we want to mimic this demo/learning app as a real app, and see
where it's edges/limits are, so we can eventually transfer some of these
learnings/lessons into real world applications.

For example, I want to be able to build a microservice architecture
using different ai-agents, as well as posibly using MCP in between. That
is the original inspiration behing this learning project in the first
place.